Problem 4
Error case: (size 2)
Error type: int is not callable
Cause: unclear understanding of expressions
" There are several atomic or primitive expressions. Numbers, booleans, strings, and the empty list (nil) are all both atomic and self-evaluating. Symbols are atomic, but are not self-evaluating (they instead evaluate to a value that was previously bound to it in the environment).

Call Expressions"

in scheme_eval(expr, env, _=None):
first check atomic, pair is not atomic
then check if is a Pair, if not Pair, Error
then check if pair.first is symbol or special form
else: call expression

Reason for making the error: